simplify, remove hacks for notify containers
